Refactored code and added some sample unit tests
I added an interface to the Six Pillars Model so that we could easily
replace out the data model with a mock one when it came time to hook
this up to a "real" data source.
I also had to increase the visibility of some properties as well as
modify the constructor of my Pillars object so that I could more easily
test it.